move some headers to private
