+++ sdk/storage/azure-storage-common/test/test_base.hpp
@@ -28,81 +28,95 @@
 #include <azure/storage/common/storage_common.hpp>
 #include <gtest/gtest.h>
 
-namespace Azure { namespace Storage { namespace Test {
+namespace Azure { namespace Storage {
 
-  const std::string& StandardStorageConnectionString();
-  const std::string& PremiumStorageConnectionString();
-  const std::string& BlobStorageConnectionString();
-  const std::string& PremiumFileConnectionString();
-  const std::string& AdlsGen2ConnectionString();
-  const std::string& AadTenantId();
-  const std::string& AadClientId();
-  const std::string& AadClientSecret();
+  namespace Test {
 
-  constexpr static const char* TestEncryptionScope = "EncryptionScopeForTest";
+    const std::string& StandardStorageConnectionString();
+    const std::string& PremiumStorageConnectionString();
+    const std::string& BlobStorageConnectionString();
+    const std::string& PremiumFileConnectionString();
+    const std::string& AdlsGen2ConnectionString();
+    const std::string& AadTenantId();
+    const std::string& AadClientId();
+    const std::string& AadClientSecret();
 
-  constexpr inline unsigned long long operator""_KB(unsigned long long x) { return x * 1024; }
-  constexpr inline unsigned long long operator""_MB(unsigned long long x)
-  {
-    return x * 1024 * 1024;
-  }
-  constexpr inline unsigned long long operator""_GB(unsigned long long x)
-  {
-    return x * 1024 * 1024 * 1024;
-  }
-  constexpr inline unsigned long long operator""_TB(unsigned long long x)
-  {
-    return x * 1024 * 1024 * 1024 * 1024;
-  }
+    constexpr static const char* TestEncryptionScope = "EncryptionScopeForTest";
 
-  std::string AppendQueryParameters(
-      const Azure::Core::Url& url,
-      const std::string& queryParameters);
+    constexpr inline unsigned long long operator""_KB(unsigned long long x) { return x * 1024; }
+    constexpr inline unsigned long long operator""_MB(unsigned long long x)
+    {
+      return x * 1024 * 1024;
+    }
+    constexpr inline unsigned long long operator""_GB(unsigned long long x)
+    {
+      return x * 1024 * 1024 * 1024;
+    }
+    constexpr inline unsigned long long operator""_TB(unsigned long long x)
+    {
+      return x * 1024 * 1024 * 1024 * 1024;
+    }
 
-  const static Azure::ETag DummyETag("0x8D83B58BDF51D75");
-  const static Azure::ETag DummyETag2("0x8D812645BFB0CDE");
-  constexpr static const char* DummyMd5 = "tQbD1aMPeB+LiPffUwFQJQ==";
-  constexpr static const char* DummyCrc64 = "+DNR5PON4EM=";
+    std::string AppendQueryParameters(
+        const Azure::Core::Url& url,
+        const std::string& queryParameters);
 
-  uint64_t RandomInt(
-      uint64_t minNumber = std::numeric_limits<uint64_t>::min(),
-      uint64_t maxNumber = std::numeric_limits<uint64_t>::max());
+    const static Azure::ETag DummyETag("0x8D83B58BDF51D75");
+    const static Azure::ETag DummyETag2("0x8D812645BFB0CDE");
+    constexpr static const char* DummyMd5 = "tQbD1aMPeB+LiPffUwFQJQ==";
+    constexpr static const char* DummyCrc64 = "+DNR5PON4EM=";
 
-  std::string RandomString(size_t size = 10);
+    uint64_t RandomInt(
+        uint64_t minNumber = std::numeric_limits<uint64_t>::min(),
+        uint64_t maxNumber = std::numeric_limits<uint64_t>::max());
 
-  std::string LowercaseRandomString(size_t size = 10);
+    std::string RandomString(size_t size = 10);
 
-  Storage::Metadata RandomMetadata(size_t size = 5);
+    std::string LowercaseRandomString(size_t size = 10);
 
-  void RandomBuffer(char* buffer, size_t length);
-  inline void RandomBuffer(uint8_t* buffer, size_t length)
-  {
-    RandomBuffer(reinterpret_cast<char*>(buffer), length);
-  }
-  std::vector<uint8_t> RandomBuffer(size_t length);
+    Storage::Metadata RandomMetadata(size_t size = 5);
 
-  inline std::vector<uint8_t> ReadBodyStream(std::unique_ptr<Azure::Core::IO::BodyStream>& stream)
-  {
-    Azure::Core::Context context;
-    return stream->ReadToEnd(context);
-  }
+    void RandomBuffer(char* buffer, size_t length);
+    inline void RandomBuffer(uint8_t* buffer, size_t length)
+    {
+      RandomBuffer(reinterpret_cast<char*>(buffer), length);
+    }
+    std::vector<uint8_t> RandomBuffer(size_t length);
 
-  inline std::vector<uint8_t> ReadBodyStream(std::unique_ptr<Azure::Core::IO::BodyStream>&& stream)
-  {
-    return ReadBodyStream(stream);
-  }
+    inline std::vector<uint8_t> ReadBodyStream(std::unique_ptr<Azure::Core::IO::BodyStream>& stream)
+    {
+      Azure::Core::Context context;
+      return stream->ReadToEnd(context);
+    }
 
-  std::vector<uint8_t> ReadFile(const std::string& filename);
+    inline std::vector<uint8_t> ReadBodyStream(
+        std::unique_ptr<Azure::Core::IO::BodyStream>&& stream)
+    {
+      return ReadBodyStream(stream);
+    }
 
-  void DeleteFile(const std::string& filename);
+    std::vector<uint8_t> ReadFile(const std::string& filename);
 
-  std::string InferSecondaryUrl(const std::string primaryUri);
+    void DeleteFile(const std::string& filename);
 
-  bool IsValidTime(const Azure::DateTime& datetime);
+    std::string InferSecondaryUrl(const std::string primaryUri);
 
-  inline std::string Base64EncodeText(const std::string& text)
-  {
-    return Azure::Core::Convert::Base64Encode(std::vector<uint8_t>(text.begin(), text.end()));
-  }
+    bool IsValidTime(const Azure::DateTime& datetime);
 
-}}} // namespace Azure::Storage::Test
+    inline std::string Base64EncodeText(const std::string& text)
+    {
+      return Azure::Core::Convert::Base64Encode(std::vector<uint8_t>(text.begin(), text.end()));
+    }
+
+  } // namespace Test
+
+  namespace Files { namespace DataLake { namespace _detail {
+
+    Azure::Core::Url GetBlobUrlFromUrl(const Azure::Core::Url& url);
+    Azure::Core::Url GetDfsUrlFromUrl(const Azure::Core::Url& url);
+    std::string GetBlobUrlFromUrl(const std::string& url);
+    std::string GetDfsUrlFromUrl(const std::string& url);
+
+  }}} // namespace Files::DataLake::_detail
+
+}} // namespace Azure::Storage
