+++ sdk/storage/azure-storage-files-datalake/CMakeLists.txt
@@ -39,13 +39,14 @@ set(
     inc/azure/storage/files/datalake/datalake_sas_builder.hpp
     inc/azure/storage/files/datalake/datalake_service_client.hpp
     inc/azure/storage/files/datalake/dll_import_export.hpp
-    inc/azure/storage/files/datalake/internal/datalake_utilities.hpp
-    inc/azure/storage/files/datalake/internal/datalake_constants.hpp
+
     inc/azure/storage/files/datalake/protocol/datalake_rest_client.hpp
 )
 
 set(
   AZURE_STORAGE_FILES_DATALAKE_SOURCE
+    src/private/datalake_constants.hpp
+    src/private/datalake_utilities.hpp
     src/private/package_version.hpp
     src/datalake_directory_client.cpp
     src/datalake_file_client.cpp
