+++ sdk/storage/azure-storage-files-datalake/src/datalake_path_client.cpp
@@ -12,9 +12,8 @@
 #include <azure/storage/common/internal/storage_switch_to_secondary_policy.hpp>
 #include <azure/storage/common/storage_common.hpp>
 
-#include "azure/storage/files/datalake/internal/datalake_constants.hpp"
-#include "azure/storage/files/datalake/internal/datalake_utilities.hpp"
-
+#include "private/datalake_constants.hpp"
+#include "private/datalake_utilities.hpp"
 #include "private/package_version.hpp"
 
 namespace Azure { namespace Storage { namespace Files { namespace DataLake {
