+++ sdk/storage/azure-storage-files-datalake/src/private/datalake_constants.hpp
