+++ sdk/storage/azure-storage-files-datalake/src/private/datalake_utilities.hpp
@@ -14,14 +14,8 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake { nam
 
   Azure::Core::Url GetBlobUrlFromUrl(const Azure::Core::Url& url);
   Azure::Core::Url GetDfsUrlFromUrl(const Azure::Core::Url& url);
-  inline std::string GetBlobUrlFromUrl(const std::string& url)
-  {
-    return GetBlobUrlFromUrl(Azure::Core::Url(url)).GetAbsoluteUrl();
-  }
-  inline std::string GetDfsUrlFromUrl(const std::string& url)
-  {
-    return GetDfsUrlFromUrl(Azure::Core::Url(url)).GetAbsoluteUrl();
-  }
+  std::string GetBlobUrlFromUrl(const std::string& url);
+  std::string GetDfsUrlFromUrl(const std::string& url);
 
   std::string SerializeMetadata(const Storage::Metadata& dataLakePropertiesMap);
