+++ sdk/storage/azure-storage-files-datalake/test/datalake_sas_test.cpp
@@ -4,7 +4,6 @@
 #include <azure/identity/client_secret_credential.hpp>
 #include <azure/storage/blobs/blob_sas_builder.hpp>
 #include <azure/storage/files/datalake/datalake_sas_builder.hpp>
-#include <azure/storage/files/datalake/internal/datalake_utilities.hpp>
 
 #include "datalake_file_system_client_test.hpp"
